requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
google-generativeai
pdfkit
pillow
wordcloud
streamlit_option_menu
pyyaml
fpdf
scipy